more MDL render progress
- add texture render modes
- do bone + chrome calculations in shader
- fix animation jitter at frame boundaries (too high epsilon for vec3 comparison)

the speed boost from doing skeletal transforms in the shader is insane. It's like 10x faster despite doing way more transformations due to the conversion to GL_TRIANGLES and the lack of shared vertex data. 50k vert transforms done in shader instead of 2k on cpu (kizuna_xmas.mdl body #20) is one especially bad case, but it's still wildly faster to do 50k transforms in the shader. Of course it depends on the GPU (I have a GTX 1070). I'm trying to keep the shader compatible with OpenGL 2.0 for the eventual port to potato PCs.

This is also what MetaHookSv does, but I remember some people saying they didn't see a speed boost from using it, so it might still be worth doing some transforms on the CPU and comparing speeds on a potato. In the case where the CPU is stronger than the GPU, doing fewer transforms on the CPU might be faster.