navigation mesh generation WIP
Ideas tried and failed:
- Using the Recast navigation library. Can't stop it from shrinking the mesh away from ledges, while still keeping it away from walls. This means narrow platforms are excluded. Invisible floors would also be missing.
- Generating from the floors of CONTENTS_EMPTY leaves. This works well for indoor areas but has polys extending past ledges. Using the intersection of empty and solid leaves had similar problems.

What's working now:
Generating the clipnodes for a hull as normal, but then cutting out all of the faces that face into the void. Many faces have to be cut. A lot of them start out inside the level then extend out into infinity outside the level. By cutting faces against each other there is a clear separation of inside and outside faces. To test if a face is facing inside the level, the contents of the BSP is checked for every point above the face (super slow). The spacing of this grid of checks has to be small or else tiny faces are excluded.

Currently the entire clipnode hull is being rendered as the mesh, and it replaces the usual hull 3 render for the world. Stripping faces that players can't walk on will make it look like a normal nav mesh.