don't autofill key values with defaults in the Smart Edit tab
if a key is not used, don't generate a value for it. If you click the field accidentally then you might add the key to the entity data without realizing it, which could possibly bug the map. Also, it should be possible to delete a key value from the smart edit tab by deleting its value.