nav mesh path finding WIP
Using dijkstra instead of AStar because some paths are extremely more expensive in order to prevent routes that require flying (not possible for players without gauss, stacking, etc.). So far it's working pretty well. It prefers to walk around the entire map and up some stairs instead of getting a bunch of players to stack in order to reach a vent in the ceiling or something.

Next major problem is having entities connect leaves together which otherwise wouldn't be, or would be too expensive due to vertical movement (teleports, ladders, water, elevators, trigger_push). Paths also sometimes take odd turns because of going to the center of a leaf before the face of the next leaf.