generate portal file for VIS compiles
adds an export option for the .prt file that VIS compilers need.

Also:
Added a render toggle for leaf connectivity for troubleshooting leaf connections. This is mostly for programmers. The leaf indexes don't match the numbering used by the VIS compiler.

Added model info for the selected entity in the debug widget.