add multi-line keyvalues and update imgui
the latest version of imgui has word wrapping for multi-line inputs, which makes them much easier to implement for keyvalues. It doesn't look great with file paths and the height calculation isn't perfect but it's better than long strings of text.

Also fixed the tooltip not wrapping for long values, and entity report scrolling with arrowkeys while not hovered.