add texture quality option
if set to low, texture atlases will be used to render the world and solid entities. This boosts FPS in the legacy renderer, but textures will shimmer in the distance due to disabled mip maps.

Texture atlases can't be used with mipmaps because of annoying borders where the textures repeat.
https://0fps.net/2013/07/09/texture-atlases-wrapping-and-mip-mapping/